Add files via upload
It is recommended to view this website on mac and full display to get the best outlook of the features, I am still working on the Media Queries to please feel free to use the Contact Page where I created a open form where users are able to send me any suggestions if needed. Note: This is not a final copy of the website, as many things are subject to change. Please stay tuned for my final product.